+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ------------------
 * Partly support for CentOS/RHEL 7
 * Note on security considerations in Readme to avoid misunderstandings
+* Show versions of intelmq and intelmq manager on about page
 
 0.2.1
 -----